requirements.txt
google_api_python_client
google_auth_oauthlib
cinemagoer
protobuf
rich
